fix: add semantic release config
This adds the semantic release config. More setting could be added, such as updating composer.json, if the plugin used that.

It will automatically create/update a changelog file with each release,
showing what changed in this release, based on the commits.

It will also automatically update the version number for the plugin.